sorry ;-) cleaned up Controllers & added adminController
