+++ app/Controllers/CvmakeController.php
@@ -11,8 +11,6 @@ $stylepick = 'cv';
 
 require 'app/Helpers/StyleSwichHelper.php';
 
-// require 'views/pages/cv-make/cv-make-home.view.php';
-
 class CVmakeController extends Controller
 {
