+++ app/Controllers/HomeController.php
@@ -2,13 +2,12 @@
 
 namespace app\Controllers;
 
+use app\Core\View;
+
 $stylepick = 'main';
 
 require 'app/Helpers/StyleSwichHelper.php';
 
-// require 'views/pages/home.view.php';
-
-use app\Core\View;
 class HomeController extends Controller
 {
 
@@ -16,4 +15,5 @@ class HomeController extends Controller
     {
         return View::render('home.view');
     }
+    
 }
