--- app/Controllers/UserController.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace app\Controllers;
-
-use app\Models\UserModel;
-use app\Core\View;
-use app\Libraries\MySql;
-class UserController extends Controller
-{
-
-    public function index()
-    {
-        $users = UserModel::all();
-    }
-
-    /**
-     * Store a user record into the database
-     */
-    public function store()
-    {
-        
-    }
-
-    /**
-     * Updates a user record into the database
-     */
-    public function update()
-    {
-        
-    }
-
-    /**
-     * Archive a user record into the database
-     */
-    public function destroy(int $id)
-    {
-
-    }
-
-}
-
