+++ routes.php
@@ -3,10 +3,11 @@
 $router->get('', 'app/Controllers/HomeController.php@index');
 $router->get('home', 'app/Controllers/HomeController.php');
 
-$router->get('logout', 'app/Controllers/LoginController.php@logout');
+$router->get('logout', 'app/Controllers/Controller.php@logout');
 
 $router->get('login', 'app/Controllers/LoginController.php');
 $router->post('login/auth', 'app/Controllers/LoginController.php@login');
+
 $router->get('invite', 'app/Controllers/InviteController.php');
 $router->post('invite/auth', 'app/Controllers/InviteController.php@login');
